pyproject.toml
[project]
name = "exchange-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiogram>=3.21.0",
    "aiohttp>=3.12.14",
    "aiomysql>=0.2.0",
    "alembic>=1.16.4",
    "asgi-lifespan>=2.1.0",
    "asyncpg>=0.30.0",
    "bcrypt[crypto]>=4.3.0",
    "fastapi>=0.116.1",
    "fastapi-mail>=1.5.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pyjwt[crypto]>=2.10.1",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "redis>=6.2.0",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.35.0",
    "websockets>=15.0.1",
]
